feat(seo): add robots.txt, fix title branding, add metadata to public pages
- Rename root title/template from "Host" to "Vardo" everywhere
- Add OG and Twitter card tags to root layout
- Add robots: noindex to authenticated (app) layout
- Create app/robots.ts — allow /login and /invite/, disallow everything else
- Extract login client logic to login-form.tsx so page.tsx can export metadata
- Add metadata export to /login (title: "Sign in")
- Add generateMetadata to /invite/[token] — dynamic org name in title and OG tags

Closes #136, closes #137
