+++ app/(app)/layout.tsx
@@ -1,9 +1,17 @@
+import type { Metadata } from "next";
 import { redirect } from "next/navigation";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import { TopNav } from "@/components/layout/top-nav";
 import { CommandPalette } from "@/components/command-palette";
 import { getCurrentOrg, getUserOrganizations } from "@/lib/auth/session";
 import { isFeatureEnabled } from "@/lib/config/features";
+
+export const metadata: Metadata = {
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
 
 export default async function AppLayout({
   children,
@@ -14,7 +22,7 @@ export default async function AppLayout({
     return (
       <div className="flex items-center justify-center min-h-dvh bg-background">
         <div className="text-center space-y-2">
-          <h1 className="text-lg font-semibold">Host</h1>
+          <h1 className="text-lg font-semibold">Vardo</h1>
           <p className="text-sm text-muted-foreground">
             Web UI is disabled. Use the API at <code className="bg-muted px-1.5 py-0.5 rounded text-xs">/api/v1/</code>
           </p>
